Show dev sign-in on SignInRequired gate
The in-page SignInRequired component (rendered on /, /instruments,
/watchers, /settings, and run detail pages for signed-out users) only
showed the Google button, so most navigation paths never surfaced the
dev sign-in affordance. Factor the dev form into a shared
DevSignInForm component used by both /login and SignInRequired so the
two gates can't drift, both keyed off the same isDevAuthEnabled flag.
